+++ core/ext.php
@@ -34,6 +34,7 @@ class ext extends \phpbb\extension\base
 		'phpbbgallery/exif',
 		'phpbbgallery/export',
 		'phpbbgallery/favorite',
+		'phpbbgallery/featured',
 		'phpbbgallery/feed',
 		'phpbbgallery/imagefields',
 		'phpbbgallery/imagerevisions',